6.02.00011 Mode 2 improvements
Attached harvesters initial support. You may need to
switch the mode to 5 and back to 2 if they don't show up
in the comine list.

Use the manual offsets from the HUD when approaching the
combine pipe. Manual offsets are relative to the automatic,
calculated offsets (left/right forward/back)

For #5317, #5342,